+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased - [0.3.0]
+
+### Added
+
+- [web-client](./web-client) MINOR create a component structure
+
 ## [0.2.0] - 2024-05-30
 
 ### Added